api/requirements.txt
flask
flask-login
flask-sqlalchemy
flask_bcrypt